Clean up some tech debt
The new restaurant page was using react-select but the new meal page
was using a classic html select so I refactored it to use react-select.
While doing that I thought it would be beneficial to reuse the
`makeOptions` function so that was moved to utils. A problem however
is that new_meal needed the value to be the id while new restaurant the
name. I couldn't just add an optional parameter to the makeOptions func
because when not passed, it would need to be index to go with the
overloaded map function provided by Javascript. Thus to keep the
new_restaurant usages of `makeOptions` still minimal, I used currying
to pass the extra variable in or not.