Attempt to fix styling of app
The last PR fixed the font awesome icon from appearing large then
resizing to the appropriate shape. Or other instances of after adding a
meal, it just staying big. Then we would have to go to a different page
and refresh and then go back to the meals index for the font to resize.

However that fix caused a different bug where many of the React
components wouldn't render with appropriate css. MUI (where most of my
components come from) has an article on using it with SSR:
https://mui.com/material-ui/guides/server-rendering/
Most of it is confusing but 1 interesting line:
"When the server receives the request, it renders the required
component(s) into an HTML string, and then sends it as a response to the
client. From that point on, the client takes over rendering duties."
Following the rest of the instructions on that page were too confusing
but I found they have examples with common React frameworks and they
had one with Remix:
https://github.com/mui/material-ui/blob/master/examples/material-ui-remix-ts/app/root.tsx
So I copied from there and modified for my use case.