Improve the layout of the header
On Kylie's phone the logout button was pushed to the side which caused
her to have to scroll sideways. Tailwind CSS has their own version of
a grid but it is more confusing than it needs to be so instead of having
to install reactstrap and potentially have SSR styling issues again, I
found MUI had a grid system similar to bootstrap so I used theirs.