Update show state of meals to show the extras
This is getting it functional as I am running out of time for this FFF.
However in the future I would like to make this UI look better. Maybe
include more info in the `AccordianSummary`. Also the main `Disclosure`
component is very similar to Accordian so there is a chance I could
change that to Accordian as well and get everything to flow nicer.