Move meals/index into a component
`useOutletContext` stopped working since the last PR where I bumped the
minor version of all the remix packages. Since they were minor bumps
I'm surprised at the breaking change. Regardless, I tried a bit to
figure out what was happening but figured this could be a traditional
React component as the loader and action could simply be in the `meals`
route. Also it was previously rendering many `<Outlet />`s which seems
bad and might have been the culprit anyhow.

The reason `eatenAt` is separated out was due to a typescript complaint
of string not being a Date.