Quit using toISOString
ISOString always sets the timezone to UTC which interferes when trying
to compare dates. Regardless, it appears all of the data in the DB has
timezone denotions. Thus I created a utility to convert to a string
and strip out all aspects of time. I should have made the data
structure to be just Dates since time is irrelevant for this feature
but this will do for now.