Fully remove toISOString usage from 75 hard
Despite what the last PR claimed to do, 75 hard was still using
toISOString() which caused the daily page to not show the check marks
if we were on it late at night. Using toDateString() returns in the
format "Thu Jan 23 2025" which normally isn't my go-to for date
comparison but since I'm using it everywhere it should be consistent.