Allow updating entries for any date
We found in use that we would sometimes forget to track (even though)
we did legit do each day. Thus setting it for other days is necessary.
This also has the benefit that if there is still a timezone issue or
something, we can go back in app to fix other days.

I kept getting errors about updating the value of an uncontrolled
component is why I had to add all the useStates and and setting in
a useEffect. This is not very clean or ideal, when I have more time
I will find a cleaner way to do this, maybe set the date in the URL
and have each date be its own route.