Change each date to be a route
That is mostly so if the user refreshes, it won't change the date on
them. The biggest fix is to go ahead and send the full datetime to the
server including timezone. Before when I was only sending MM/DD/YYYY,
the server would be able to apply its own timezone. But if the client
sends its timezone, then the server has to use that. Dayjs needs an
extension to use timezones so I figured it would be simpler to convert
from dayjs -> Date -> to string with timezone since the built in JS Date
object automatically uses timezone.

I also cleaned up the console errors by not allowing the input value to
ever be `undefined` by instead defaulting to empty string or false if
there is no entry for a day. I think to clean the code up in the future
would be to auto create an entry in the loader if it doesn't exist then
the frontend will always have an entry and not need to worry about
defaults.

Also I think having controlled components is a must for remix due to
however hydration works, it doesn't actually fully re-render components
even when the date route changes.