Dont let entry be null on the frontend
Now in the loader (on the server) I check if there is an Entry record
and if not, create it. At that point since I would for sure have an
Entry, I figured I might as well have the loader return that and then
I wouldn't need to find the entry with the matching date (on the client).
This is huge because it moved all date comparisons to the backend. I
was getting nervous about timezone differences but then after more
digging I noticed the dates in the DB are already stored as ISO8601 and
when I console logged `entry.date`, it outputted the ISO string. I
further verified it in the prisma docs:
https://www.prisma.io/docs/orm/prisma-client/special-fields-and-types#working-with-datetime
I then dug more into the dayjs docs and its default string parsing
constructor expects the string to be in ISO8601:
https://day.js.org/docs/en/parse/string and there is even a warning box
in there that if you pass a string that is not ISO8601 then there might
be inconsistencies... whoops. Therefore despite previous attempts to
fully get away from ISO strings, I have now come back full circle to
embrace it. I will keep the URLSearchParams as `MM-DD-YYYY` to keep that
cleaner as well as since once again, this app really doesn't even care
about times and timezones. Therefore, keeping the param in that format
also makes it so when an entry is created, it won't be setting a time.