Make the Calendar consistent with the Day page
As explained in the previous commit, the dates are already stored in ISO
format. That can be verified by console logging the challenge.startDate.
Now that dayjs is "safe" to use again, it does clean things up and makes
it easier to do things like compare dates, add, and format.