Fix Date parsing error
I think I finally figured it out. According the JS Date docs:
```
When the time zone offset is absent, date-only forms are interpreted as
a UTC time and date-time forms are interpreted as a local time.
```
Thus by me trying to take off the time and just use a string with month
day and year, I was actually hurting myself because then the Date obj
stored in the DB was purely in UTC rather than being converted from the
local timezone to UTC. dayjs and the DatePicker component will then read
the DateTime in UTC and convert it to the local.